# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AcmeU Platform API
  version: "1.0"
  description: |
    Public API for the AcmeU learning and credentialing platform.

    All endpoints are tenant-scoped. Access tokens are RS256 JWTs issued by
    `POST /auth/login` and verified against the JWKS document published at
    `/.well-known/jwks.json`. Service integrations may authenticate with a
    scoped API key supplied in the `X-API-Key` header.
  contact:
    name: AcmeU Platform Team
    email: platform@acmeu.com
servers:
  - url: https://api.acmeu.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Exchange credentials for an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                otp: { type: string, description: Required when the account has MFA enabled }
                rememberMe: { type: boolean, default: false }
                tenant: { type: string, description: Workspace slug, when the email exists in more than one workspace }
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/register:
    post:
      tags: [Authentication]
      summary: Create a learner account in an existing workspace
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName, tenantSlug]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
                displayName: { type: string }
                tenantSlug: { type: string }
                inviteToken: { type: string }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /courses/{courseId}/assignments:
    get:
      tags: [Assessment]
      summary: Assignments for a course, with the caller's submission state
      parameters: [{ $ref: "#/components/parameters/CourseId" }]
      responses:
        "200": { description: Assignment list }
  /courses/{courseId}/submissions/{submissionId}:
    get:
      tags: [Assessment]
      summary: Retrieve a submission
      parameters:
        - { $ref: "#/components/parameters/CourseId" }
        - { in: path, name: submissionId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Submission detail }
        "404": { $ref: "#/components/responses/NotFound" }
  /submissions/{submissionId}/grade:
    post:
      tags: [Assessment]
      summary: Record a grade against a submission
      parameters:
        - { in: path, name: submissionId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Graded }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    CourseId:
      in: path
      name: courseId
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request payload failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication is required or the credential was rejected
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The caller lacks the required role or scope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
            requestId: { type: string }
    Session:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string }
            displayName: { type: string }
            role: { type: string, enum: [learner, instructor, manager, admin, owner] }
            tenantId: { type: string, format: uuid }
...